Full pipeline

+++ crow/crow/entrypoints/bc2wasm.py
@@ -1,16 +1,14 @@
 
 from crow.commands.stages import ObjtoWASM
-from crow.events import LOG_MESSAGE, C2LL_MESSAGE, LL2BC_MESSAGE, BC2Candidates_MESSAGE, BC2WASM_MESSAGE, STORE_MESSAGE, WASM_QUEUE, WASM2WAT_MESSAGE
+from crow.events import BC2WASM_MESSAGE, STORE_MESSAGE, WASM_QUEUE, WASM2WAT_MESSAGE
 from crow.events.event_manager import Publisher, Subscriber, subscriber_function
-from crow.logger import ERROR
 from crow.settings import config
 
 import sys
-from crow.utils import printProgressBar, createTmpFile, getIteratorByName, \
-    ContentToTmpFile, BreakException, RUNTIME_CONFIG, updatesettings, NOW
+from crow.utils import ContentToTmpFile
 
 import os
-from crow.monitor.logger import log_system_exception
+from crow.monitor.monitor import log_system_exception
 
 @log_system_exception()
 def bc2wasm(bc, program_name, file_name=None):
@@ -19,34 +17,37 @@ def bc2wasm(bc, program_name, file_name=None):
 
     file_name = program_name if file_name is None else file_name
 
-    print(file_name)
-    with ContentToTmpFile(name="%s.bc" % file_name, content=bc, ext=".bc", persist=False) as TMP_WASM:
-        tmpWasm = TMP_WASM.file
 
-        finalObjCreator = ObjtoWASM(file_name, debug=True)
-        finalObjCreator(args=[
-            "%s.wasm" % (file_name,),
-            tmpWasm
-        ], std=None)
+    # 2 WASM transformation
+    if config["DEFAULT"].getboolean("keep-wasm-files"):
+        with ContentToTmpFile(name="%s.bc" % file_name, content=bc, ext=".bc", persist=False) as TMP_WASM:
+            tmpWasm = TMP_WASM.file
 
-        # Explicitly saving wasm file
-        publisher.publish(message=dict(
-            event_type=STORE_MESSAGE,
-            stream=open(f"{file_name}.wasm", 'rb').read(),
-            program_name=program_name,
-            file_name=f"{file_name}.wasm"
-        ), routing_key="")
+            finalObjCreator = ObjtoWASM(file_name, debug=True)
+            finalObjCreator(args=[
+                "%s.wasm" % (file_name,),
+                tmpWasm
+            ], std=None)
+
+            # Explicitly saving wasm file
+            publisher.publish(message=dict(
+                event_type=STORE_MESSAGE,
+                stream=open(f"{file_name}.wasm", 'rb').read(),
+                program_name=program_name,
+                file_name=f"{file_name}.wasm"
+            ), routing_key="")
 
 
-        # Generate wat file ?
-        publisher.publish(message=dict(
-            event_type=WASM2WAT_MESSAGE,
-            stream=open(f"{file_name}.wasm", 'rb').read(),
-            program_name=program_name,
-            file_name=f"{file_name}"
-        ), routing_key="")
+            # Generate wat file ?
+            if config["DEFAULT"].getboolean("keep-wat-files"):
+                publisher.publish(message=dict(
+                    event_type=WASM2WAT_MESSAGE,
+                    stream=open(f"{file_name}.wasm", 'rb').read(),
+                    program_name=program_name,
+                    file_name=f"{file_name}"
+                ), routing_key="")
 
-        os.remove("%s.wasm" % (file_name,))
+            os.remove("%s.wasm" % (file_name,))
 
 
 @log_system_exception()
